refactor: Extract Pydantic schemas to kv_schemas.py
Move all request/response schemas from kv_router.py to a dedicated
kv_schemas.py module for better separation of concerns.

- Creates automation/kv_schemas.py with 12 Pydantic models
- Updates kv_router.py to import from kv_schemas
- Reduces kv_router.py from 969 to 883 lines
